==> code/05-pipeline.qmd <==
---
title: "Augmentation-Pipeline"
format: html
---

## Purpose

Prepare dataset for modelling

```{r}
library(tidyverse)
library(glue)
library(janitor)
library(glue)

files <- list.files(path = "data/raw", pattern = ".rds") %>% str_c("data/raw/", .)

df_list <- map(files, read_rds)
df <- bind_rows(df_list, .id = 'date_num')

df <- df %>% 
  clean_names()

df <- df %>% 
  mutate(across(c(year, kilometers), parse_number))

df <- df %>%
  mutate(make_model = glue("{make} {model}"),
         price = round(price),
         year = round(year)) %>%
  select(title,
         price,
         year,
         kilometers,
         make_model,
         seller_type,
         seller_name,
         province,
         location,
         location_2,
         colour,
         n_photos,
         body_type,
         fuel_type,
         transmission,
         ad_url,
         ad_date,
         n_views,
         text)

df <- df %>% 
  mutate(province = str_replace(province, "-", " "),
         province = str_replace(province, "\\+", "-"),
         province = str_to_title(province))

df <- df %>%
  arrange(ad_url, desc(ad_date)) %>%
  distinct(title, kilometers, text, .keep_all = T)

```

Join to location data

```{r}
locations <- read_rds("data/clean/locations.rds") %>% 
  rename(location = location_key) %>% 
  mutate(location = str_squish(str_remove(location, ", South Africa")))

df <- df %>% 
  left_join(locations, by = c("location_2" = "location")) %>% 
  select(-location) %>% 
  rename(location = location_2)

```

What about some data sanity checks??

For instance, year should be truncated to like 1990, no?

Maybe I should make a document that records these choices? Or put it here?

### Data sanitisation

```{r}
# df %>%
#   ggplot(aes(year)) +
#   geom_density() +
#   geom_vline(xintercept = 1990, lty = 2)
# 
# df %>%
#   ggplot(aes(price)) +
#   geom_density() +
#   scale_x_continuous(limits = c(NA, 2000000))
# 
# df %>% 
#   ggplot(aes(kilometers)) +
#   geom_density() +
#   scale_x_continuous(limits = c(NA, 500000))

df <- df %>% 
  filter(year >= 1990,
         price <= 2000000,
         kilometers < 500000)

st <- format(Sys.time(), "%Y-%m-%d-%I-%M-%p")

write_rds(df, paste0("clean_data/df_complete_", st, ".rds"), compress = "gz")
```
